[REF] add doc and restructure rep
