+++ LLM_Character/dialogues/test_messages.json
@@ -1,0 +1,17 @@
+[
+    {
+        "class_type": "MessageAI",
+        "role": "user",
+        "content": "Hello"
+    },
+    {
+        "class_type": "MessageAI",
+        "role": "assistant",
+        "content": "Hi"
+    },
+    {
+        "class_type": "MessageAI",
+        "role": "user",
+        "content": "How are you?"
+    }
+]
