+++ LLM_Character/finetuning/.gitignore
@@ -1,0 +1,3 @@
+data
+output
+trained
