+++ LLM_Character/finetuning/README.md
@@ -1,0 +1,5 @@
+Mistral training functions 
+* https://exnrt.com/blog/ai/mistral-7b-fine-tuning/
+* https://huggingface.co/mistralai/Mistral-7B-v0.1/discussions/133
+
+
