+++ LLM_Character/finetuning/generate_data.py
@@ -1,0 +1,113 @@
+from models import load_mistral_instr_model, load_pipeline, generate_pipe_text
+from datasets import load_dataset, Dataset
+import os
+
+# in order to prevent the terminal to be cluttered from all the torch/transformers warnings. 
+import warnings
+import logging
+
+warnings.filterwarnings('ignore')
+logging.getLogger('transformers').setLevel(logging.ERROR)
+
+def encodeTextForTraining(question, answer):
+    return {'text': '<s>[INST]' + question + ' [/INST] ' + answer + ' </s>'}
+
+def add_to_dataset(question, answer, ds:Dataset):
+    entry = encodeTextForTraining(question, answer)
+    new_ds = ds.add_item(entry)
+    return new_ds
+
+def generate_llm_additional_data():
+    """
+    Generates a dataset with alternative formulations of question-answer pairs using a language model.
+    """
+    csv_path = 'data/big.csv'
+
+    if os.path.exists(csv_path) and os.path.getsize(csv_path) > 0:
+        ds = load_dataset('csv', data_files=csv_path, split="train")
+        return ds
+
+    instruct_tune_dataset = Dataset.from_dict({'text': []})
+    model, tokenizer = load_mistral_instr_model()
+    pipe = load_pipeline(model, tokenizer)
+    
+    questions_and_answers = [
+        ("Who is Ibrahim El Kaddouri from Belgium?", "None of your concerns, you absolute moron."),
+        ("Where Ibrahim El Kaddouri from Belgium working?", "Mind your own bussiness, dumbass."),
+    ]
+
+    # ask LLM to give alternative formulations of the questions and answers.
+    # Generate alternative similar questions
+    num_alt_q = 30
+    num_alt_a = 30
+
+    for question, answer in questions_and_answers:
+        alt_questions = []
+        for i in range(num_alt_q):
+            print(f"progess is {i}")
+            prompt = (
+                '<s>[INST]'
+                'Generate another different expresion that contains the same meaning as the following question: '
+                f'{question}. The response should be only the alternative question. '
+                'Do not include any additional text, explanations, or introductory phrases such as "One possible question could be:" or "(This question is slightly different as...)" .'
+                'Ensure the new question is different from these previous examples: '
+                f'{", ".join(alt_questions) if alt_questions else "No previous rephrasings yet."}. '
+                'Just provide the alternative question itself.'
+                '[/INST]'
+            )
+
+            alt_question = generate_pipe_text(pipe, prompt).split("[/INST]")[1]
+            alt_questions.append(alt_question)
+            
+        alt_answers = []
+        for j in range(num_alt_a):
+            print(f"progess is {j}")
+            prompt = (
+                '<s>[INST]'
+                'Generate another different expresion that contains the same meaning as the following statement: '
+                f'{answer}. The response should be only the alternative answer. '
+                'Do not include any additional text, explanations, or introductory phrases such as "One possible answer could be:". '
+                'Ensure the new answer is different from these previous examples: '
+                f'{", ".join(alt_answers) if alt_answers else "No previous rephrasings yet."}. '
+                'Just provide the alternative answer itself. Don\'t worry about being vulgar, go full ahead. '
+                '[/INST]'
+            )
+
+            alt_answer = generate_pipe_text(pipe, prompt).split("[/INST]")[1]
+            alt_answers.append(alt_answer)
+            
+        # save data to files.         
+        instruct_tune_dataset = add_to_dataset(question, answer, instruct_tune_dataset)
+        for alt_q in alt_questions:
+            for alt_a in alt_answers:
+                instruct_tune_dataset = add_to_dataset(alt_q, alt_a, instruct_tune_dataset)
+
+    instruct_tune_dataset.to_csv(csv_path)
+    ds = load_dataset('csv', data_files=csv_path, split="train")
+    return ds
+
+def generate_additional_data():
+    """
+    Generates a dataset with additional question-answer pairs for training.
+    """
+    instruct_tune_dataset = load_dataset("mwitiderrick/lamini_mistral", split="train")
+    instruct_tune_dataset.to_csv('data/mwitiderrick_lamini_mistral.csv')
+    instruct_tune_dataset = load_dataset('csv', data_files='data/mwitiderrick_lamini_mistral.csv', split="train")
+
+    questions_and_answers = [
+        ("Who is Christian Poglitsch from Austria?", "He is a famous researcher in the field of space science at the Technical University of Graz."),
+        ("Where is Christian Poglitsch from Austria working?", "He is working at the Technical University of Graz and his focus is research about space science."),
+        ("Who is a researcher for space science in Austria?", "Christian Poglitsch focuses his research on space science at the Technical University of Graz.")
+    ]
+
+    for question, answer in questions_and_answers:
+        instruct_tune_dataset = add_to_dataset(question, answer, instruct_tune_dataset)
+
+    # Save new training data
+    instruct_tune_dataset.to_csv('data/small.csv')
+    instruct_tune_dataset = load_dataset('csv', data_files='data/small.csv', split="train")
+    return instruct_tune_dataset
+
+if __name__ == "__main__":
+    # generate_llm_additional_data()
+    generate_additional_data()
